+++ Libplanet/ByteUtil.cs
@@ -5,6 +5,7 @@ using System.Collections.Immutable;
 using System.Diagnostics.Contracts;
 using System.Globalization;
 using System.Linq;
+using System.Numerics;
 
 namespace Libplanet
 {
@@ -138,6 +139,47 @@ namespace Libplanet
             }
 
             return !differ;
+        }
+
+        /// <summary>
+        /// Tests if a hash digest is less than the target computed for the given
+        /// <paramref name="difficulty"/>).
+        /// </summary>
+        /// <param name="hashDigest">A hash digest to test.</param>
+        /// <param name="difficulty">The difficulty to compute target number.</param>
+        /// <returns><c>true</c> only if a digest is less than the target computed for the given
+        /// <paramref name="difficulty"/>).  If <paramref name="difficulty"/> is <c>0</c> it always
+        /// returns <c>true</c>.</returns>
+        [Pure]
+        public static bool Satisfies(IReadOnlyList<byte> hashDigest, long difficulty)
+        {
+            if (difficulty == 0)
+            {
+                return true;
+            }
+            else if (!hashDigest.Any())
+            {
+                return false;
+            }
+
+            var maxTargetBytes = new byte[hashDigest.Count + 1];
+            maxTargetBytes[hashDigest.Count] = 0x01;
+            var maxTarget = new BigInteger(maxTargetBytes);
+            BigInteger target = maxTarget / difficulty;
+
+            var digestArray = new byte[hashDigest.Count + 1];
+            int i = 0;
+            foreach (byte b in hashDigest)
+            {
+                digestArray[i++] = b;
+            }
+
+            // Append zero to convert unsigned BigInteger.  Note that BigInteger(byte[]) assumes
+            // the input bytes are in little-endian order.
+            digestArray[i] = 0;
+
+            var result = new BigInteger(digestArray);
+            return result < target;
         }
     }
 }